+++ calendar-bundle/contao/config/config.php
@@ -11,7 +11,7 @@
 use Contao\CalendarEventsModel;
 use Contao\CalendarFeedModel;
 use Contao\CalendarModel;
-use Contao\CoreBundle\Controller\BackendCsvImportController;
+use Contao\CoreBundle\Controller\Backend\CsvImportController;
 use Contao\ModuleCalendar;
 use Contao\ModuleEventlist;
 use Contao\ModuleEventMenu;
@@ -23,8 +23,8 @@ use Symfony\Component\HttpFoundation\Request;
 $GLOBALS['BE_MOD']['content']['calendar'] = array
 (
 	'tables'      => array('tl_calendar', 'tl_calendar_events', 'tl_content'),
-	'table'       => array(BackendCsvImportController::class, 'importTableWizardAction'),
-	'list'        => array(BackendCsvImportController::class, 'importListWizardAction')
+	'table'       => array(CsvImportController::class, 'importTableWizardAction'),
+	'list'        => array(CsvImportController::class, 'importListWizardAction')
 );
 
 // Front end modules
