+++ core-bundle/contao/config/config.php
@@ -22,7 +22,7 @@ use Contao\ContentModel;
 use Contao\ContentModule;
 use Contao\ContentSliderStart;
 use Contao\ContentSliderStop;
-use Contao\CoreBundle\Controller\BackendCsvImportController;
+use Contao\CoreBundle\Controller\Backend\CsvImportController;
 use Contao\Crawl;
 use Contao\FilesModel;
 use Contao\FileTree;
@@ -125,8 +125,8 @@ $GLOBALS['BE_MOD'] = array
 		(
 			'tables'      => array('tl_article', 'tl_content'),
 			'ptables'     => array('tl_page'),
-			'table'       => array(BackendCsvImportController::class, 'importTableWizardAction'),
-			'list'        => array(BackendCsvImportController::class, 'importListWizardAction')
+			'table'       => array(CsvImportController::class, 'importTableWizardAction'),
+			'list'        => array(CsvImportController::class, 'importListWizardAction')
 		),
 		'files' => array
 		(
@@ -135,7 +135,7 @@ $GLOBALS['BE_MOD'] = array
 		'form' => array
 		(
 			'tables'      => array('tl_form', 'tl_form_field'),
-			'option'      => array(BackendCsvImportController::class, 'importOptionWizardAction')
+			'option'      => array(CsvImportController::class, 'importOptionWizardAction')
 		)
 	),
