+++ core-bundle/src/Controller/Backend/CsvImportController.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
  * @license LGPL-3.0-or-later
  */
 
-namespace Contao\CoreBundle\Controller;
+namespace Contao\CoreBundle\Controller\Backend;
 
 use Contao\BackendTemplate;
 use Contao\Config;
@@ -27,7 +27,7 @@ use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
-class BackendCsvImportController
+class CsvImportController
 {
     final public const SEPARATOR_COMMA = 'comma';
