+++ news-bundle/contao/config/config.php
@@ -8,7 +8,7 @@
  * @license LGPL-3.0-or-later
  */
 
-use Contao\CoreBundle\Controller\BackendCsvImportController;
+use Contao\CoreBundle\Controller\Backend\CsvImportController;
 use Contao\ModuleNewsArchive;
 use Contao\ModuleNewsList;
 use Contao\ModuleNewsMenu;
@@ -22,8 +22,8 @@ use Symfony\Component\HttpFoundation\Request;
 $GLOBALS['BE_MOD']['content']['news'] = array
 (
 	'tables'      => array('tl_news_archive', 'tl_news', 'tl_content'),
-	'table'       => array(BackendCsvImportController::class, 'importTableWizardAction'),
-	'list'        => array(BackendCsvImportController::class, 'importListWizardAction')
+	'table'       => array(CsvImportController::class, 'importTableWizardAction'),
+	'list'        => array(CsvImportController::class, 'importListWizardAction')
 );
 
 // Front end modules
